google strategy also refactored into middleware
