+++ backend/src/api/index.ts
@@ -13,29 +13,12 @@ import setupSwaggerUI from './apiDocumentation'
 import { createServiceLogger } from '../utils/logging'
 import { responseHandlerMiddleware } from '../middlewares/responseHandlerMiddleware'
 import { errorMiddleware } from '../middlewares/errorMiddleware'
-import { createRedisClient, RedisCache } from '../utils/redis'
 import { passportStrategyMiddleware } from '../middlewares/passportStrategyMiddleware'
 import { redisMiddleware } from '../middlewares/redisMiddleware'
 
 const serviceLogger = createServiceLogger()
 
 const app = express()
-
-setImmediate(async () => {
-  const client = await createRedisClient(true)
-  await client.ping()
-
-  const cache = new RedisCache('test', client)
-
-  await cache.setValue('blabla', '123')
-  const value = await cache.getValue('blabla')
-  console.log(value)
-
-  console.log(await cache.getValueByKeyPattern('bla*'))
-  await cache.deleteByKeyPattern('bla*')
-
-  console.log(await cache.getValue('blabla'))
-})
 
 // Enables CORS
 app.use(cors({ origin: true }))
